refactor(daemon): isolate standing query stage
Problem: the convergence-stage retro requires extracting a stage before\nadding a fourth one, and multiple watched aliases could evaluate the same\nimmutable definition twice in a single convergence pass.\n\nWhat changed: move standing-query materialization into its own daemon\nmodule while retaining the existing factory export, and deduplicate watched\nquery versions at the durable lookup boundary. Add a regression covering\ntwo aliases resolving to one evaluation and no fabricated initial delta.\n\nRef polylogue-rxdo.5\nRef polylogue-rxdo.6\n\nCo-Authored-By: Claude <noreply@anthropic.com>
