+++ docs/topology-status.md
@@ -28,12 +28,12 @@ Generated by `devtools render topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 774
+- **Stable** (no move scoped): 775
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 18
 - **TBD** (cell needs explicit assignment): 6
-- **Total declared**: 928
-- **Realized polylogue/**/*.py**: 928 files declared
+- **Total declared**: 929
+- **Realized polylogue/**/*.py**: 929 files declared
 
 ### TBD cells (require explicit routing)
